Lyrics, RandomByLastfm, Telepathy: Update datatype of DB IDs
Commit f847cb0c742b2d993dad71dba373bd81bb457afa changed the handling of database
IDs from int to long, so this needs to be changed appropriately here as well.

Fixes BGO#705422

+++ src/Telepathy/Banshee.Telepathy/Banshee.Telepathy.DBus/MetadataProviderService.cs
@@ -121,7 +121,7 @@ namespace Banshee.Telepathy.DBus
         public int[] GetPlaylistIds (LibraryType type)
         {
             //Console.WriteLine ("I am in GetPlaylistIds");
-            int primary_source_id = 0;
+            long primary_source_id = 0;
 
             switch (type) {
                 case LibraryType.Music:
